backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
cryptography==41.0.7
openpyxl==3.1.2
python-decouple==3.8
anthropic==0.7.8
openai==1.3.7
django-extensions==3.2.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
pandas==2.1.3
xlsxwriter==3.1.9